+++ packages/agent/src/actor.ts
@@ -1,10 +1,9 @@
-import { Buffer } from 'buffer/';
 import {
   Agent,
   HttpDetailsResponse,
+  isV2ResponseBody,
+  isV3ResponseBody,
   QueryResponseStatus,
-  v2ResponseBody,
-  v3ResponseBody,
 } from './agent';
 import {
   CertifiedRejectErrorCode,
@@ -18,8 +17,9 @@ import {
   UnknownError,
 } from './errors';
 import { bufFromBufLike, IDL } from '@dfinity/candid';
-import { pollForResponse, PollStrategyFactory, strategy } from './polling';
+import { pollForResponse, PollingOptions, DEFAULT_POLLING_OPTIONS } from './polling';
 import { Principal } from '@dfinity/principal';
+import { strToUtf8 } from './utils/buffer';
 import { Certificate, CreateCertificateOptions, lookupResultToBuffer } from './certificate';
 import managementCanisterIdl from './canisters/management_idl';
 import _SERVICE, { canister_install_mode, canister_settings } from './canisters/management_service';
@@ -36,10 +36,9 @@ export interface CallConfig {
   agent?: Agent;
 
   /**
-   * A polling strategy factory that dictates how much and often we should poll the
-   * read_state endpoint to get the result of an update call.
+   * Options for controlling polling behavior.
    */
-  pollingStrategyFactory?: PollStrategyFactory;
+  pollingOptions?: PollingOptions;
 
   /**
    * The canister ID of this Actor.
@@ -83,6 +82,11 @@ export interface ActorConfig extends CallConfig {
    * Polyfill for BLS Certificate verification in case wasm is not supported
    */
   blsVerify?: CreateCertificateOptions['blsVerify'];
+
+  /**
+   * Polling options to use when making update calls. This will override the default DEFAULT_POLLING_OPTIONS.
+   */
+  pollingOptions?: PollingOptions;
 }
 
 // TODO: move this to proper typing when Candid support TypeScript.
@@ -385,7 +389,7 @@ export class Actor {
 // produces an array. Ensure that functions with single or zero return
 // values behave as expected.
 function decodeReturnValue(types: IDL.Type[], msg: ArrayBuffer) {
-  const returnValues = IDL.decode(types, Buffer.from(msg));
+  const returnValues = IDL.decode(types, msg);
   switch (returnValues.length) {
     case 0:
       return undefined;
@@ -396,8 +400,8 @@ function decodeReturnValue(types: IDL.Type[], msg: ArrayBuffer) {
   }
 }
 
-const DEFAULT_ACTOR_CONFIG = {
-  pollingStrategyFactory: strategy.defaultStrategy,
+const DEFAULT_ACTOR_CONFIG: Partial<ActorConfig> = {
+  pollingOptions: DEFAULT_POLLING_OPTIONS,
 };
 
 export type ActorConstructor = new (config: ActorConfig) => ActorSubclass;
@@ -474,7 +478,8 @@ function _createActorMethod(
       };
 
       const agent = options.agent || actor[metadataSymbol].config.agent || HttpAgent.createSync();
-      const { canisterId, effectiveCanisterId, pollingStrategyFactory } = {
+
+      const { canisterId, effectiveCanisterId, pollingOptions } = {
         ...DEFAULT_ACTOR_CONFIG,
         ...actor[metadataSymbol].config,
         ...options,
@@ -490,18 +495,18 @@ function _createActorMethod(
       });
       let reply: ArrayBuffer | undefined;
       let certificate: Certificate | undefined;
-      if (response.body && (response.body as v3ResponseBody).certificate) {
+      if (isV3ResponseBody(response.body)) {
         if (agent.rootKey == null) {
           throw ExternalError.fromCode(new MissingRootKeyErrorCode());
         }
-        const cert = (response.body as v3ResponseBody).certificate;
+        const cert = response.body.certificate;
         certificate = await Certificate.create({
           certificate: bufFromBufLike(cert),
           rootKey: agent.rootKey,
           canisterId: Principal.from(canisterId),
           blsVerify,
         });
-        const path = [new TextEncoder().encode('request_status'), requestId];
+        const path = [strToUtf8('request_status'), requestId];
         const status = new TextDecoder().decode(
           lookupResultToBuffer(certificate.lookup([...path, 'status'])),
         );
@@ -518,6 +523,7 @@ function _createActorMethod(
             const rejectMessage = new TextDecoder().decode(
               lookupResultToBuffer(certificate.lookup([...path, 'reject_message']))!,
             );
+
             const error_code_buf = lookupResultToBuffer(
               certificate.lookup([...path, 'error_code']),
             );
@@ -533,9 +539,9 @@ function _createActorMethod(
             );
           }
         }
-      } else if (response.body && 'reject_message' in response.body) {
+      } else if (isV2ResponseBody(response.body)) {
         // handle v2 response errors by throwing an UpdateCallRejectedError object
-        const { reject_code, reject_message, error_code } = response.body as v2ResponseBody;
+        const { reject_code, reject_message, error_code } = response.body;
         throw RejectError.fromCode(
           new CertifiedRejectErrorCode(requestId, reject_code, reject_message, error_code, {
             canisterId: cid,
@@ -547,9 +553,12 @@ function _createActorMethod(
 
       // Fall back to polling if we receive an Accepted response code
       if (response.status === 202) {
-        const pollStrategy = pollingStrategyFactory();
+        const pollOptions: PollingOptions = {
+          ...pollingOptions,
+          blsVerify,
+        };
         // Contains the certificate and the reply from the boundary node
-        const response = await pollForResponse(agent, ecid, requestId, pollStrategy, blsVerify);
+        const response = await pollForResponse(agent, ecid, requestId, pollOptions);
         certificate = response.certificate;
         reply = response.reply;
       }