+++ packages/agent/src/canisterStatus/index.test.ts
@@ -5,7 +5,7 @@ import { fromHexString } from '@dfinity/candid';
 import { Identity } from '../auth';
 import fetch from 'isomorphic-fetch';
 import { HttpAgent } from '../agent';
-import { fromHex, toHex } from '../utils/buffer';
+import { bufFromBufLike, fromHex, strToUtf8, toHex } from '../utils/buffer';
 import * as Cert from '../certificate';
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-ignore
@@ -28,13 +28,7 @@ jest.mock('../utils/bls', () => {
 jest.useFakeTimers();
 const certificateTime = Date.parse('2022-05-19T20:58:22.596Z');
 jest.setSystemTime(certificateTime);
-
-// Utils
-const encoder = new TextEncoder();
-const encode = (arg: string): ArrayBuffer => {
-  return new DataView(encoder.encode(arg).buffer).buffer;
-};
-const canisterBuffer = new DataView(testPrincipal.toUint8Array().buffer).buffer;
+const canisterBuffer = bufFromBufLike(testPrincipal.toUint8Array());
 
 /* Produced by deploying a dfx new canister and requesting
   | 'time'
@@ -62,19 +56,18 @@ const getRealStatus = async () => {
 
   const agent = new HttpAgent({ host: 'http://127.0.0.1:4943', fetch, identity });
   await agent.fetchRootKey();
-  const canisterBuffer = new DataView(testPrincipal.toUint8Array().buffer).buffer;
+  const canisterBuffer = bufFromBufLike(testPrincipal.toUint8Array().buffer);
   const response = await agent.readState(
     testPrincipal,
     // Note: subnet is not currently working due to a bug
     {
       paths: [
         encodePath('time', testPrincipal),
-        [encode('canister'), canisterBuffer, encode('controllers')],
-        [encode('canister'), canisterBuffer, encode('module_hash')],
+        [strToUtf8('canister'), canisterBuffer, strToUtf8('controllers')],
+        [strToUtf8('canister'), canisterBuffer, strToUtf8('module_hash')],
         encodePath('candid', testPrincipal),
       ],
     },
-    identity,
   );
   console.log(toHex(response.certificate));
 };
@@ -115,14 +108,14 @@ describe('Canister Status utility', () => {
     const status = await getStatus([
       {
         key: 'time',
-        path: [new DataView(new TextEncoder().encode('time').buffer).buffer],
+        path: [strToUtf8('time')],
         decodeStrategy: 'leb128',
       },
     ]);
     const statusRaw = await getStatus([
       {
         key: 'time',
-        path: [new DataView(new TextEncoder().encode('time').buffer).buffer],
+        path: [strToUtf8('time')],
         decodeStrategy: 'raw',
       },
     ]);
@@ -137,14 +130,14 @@ describe('Canister Status utility', () => {
     const statusHex = await getStatus([
       {
         key: 'time',
-        path: [new DataView(new TextEncoder().encode('time').buffer).buffer],
+        path: [strToUtf8('time')],
         decodeStrategy: 'hex',
       },
     ]);
     const statusCBOR = await getStatus([
       {
         key: 'Controller',
-        path: [encode('canister'), canisterBuffer, encode('controllers')],
+        path: [strToUtf8('canister'), canisterBuffer, strToUtf8('controllers')],
         decodeStrategy: 'cbor',
       },
     ]);
@@ -166,7 +159,7 @@ describe('Canister Status utility', () => {
     const statusEncoded = await getStatus([
       {
         kind: 'metadata',
-        path: encode('candid:service'),
+        path: strToUtf8('candid:service'),
         key: 'candid',
         decodeStrategy: 'hex',
       },
@@ -188,7 +181,7 @@ describe('Canister Status utility', () => {
       'subnet',
       {
         key: 'asdf',
-        path: [new DataView(new TextEncoder().encode('asdf').buffer).buffer],
+        path: [strToUtf8('asdf')],
         decodeStrategy: 'hex',
       },
     ]);